backend/requirements.txt
Flask
Werkzeug
flask-cors
python-dotenv
pydantic
numpy
pandas
scikit-learn
xgboost
scipy
joblib
PyMuPDF
pdfplumber
python-docx
beautifulsoup4
lxml
filetype
chardet
regex
Unidecode
dateparser
rapidfuzz
openai